Re-factored the NotificationsTest to pull out `StubOpportunityOpps` into a trait
